requirements.txt
alembic==1.5.5
asgiref==3.4.1
astroid==2.6.2
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
colorama==0.4.4
deepdiff==5.2.3
defusedxml==0.7.1
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
httpie==2.5.0
idna==3.2
isort==5.9.2
jsonpickle==2.0.0
lazy-object-proxy==1.6.0
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
mergedeep==1.3.4
numpy==1.19.3
ordered-set==4.0.2
pandas==1.3.2
prettytable==2.1.0
psutil==5.8.0
pycodestyle==2.7.0
Pygments==2.10.0
PyJWT==2.1.0
pylint==2.9.3
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23
sqlparse==0.4.2
toml==0.10.2
tqdm==4.57.0
ujson==4.0.2
urllib3==1.26.7
wcwidth==0.2.5
win32-setctime==1.0.3
wrapt==1.12.1
